- (Diep) - Cleaned up Quadrivium files.
