+++ src/app/app.component.html
@@ -20,7 +20,7 @@
                 <h1 #heyThere (click)="shakeHeyThere()" style="cursor:pointer;">Hey there, friend.</h1>
                 <p #goodNews (click)="shakeGoodNews()" style="cursor:pointer;">
                     Good news! My angular website is running, customized,
-                    <br> and all links are updated as of 2026.05.29! 🎉
+                    <br> and all links are updated as of 2026.06.01! 🎉
                 </p>
 
                 <div class="view-selector-container">
